Triage remaining static analysis findings; type model relations
- Added explicit return types to all 26 Eloquent relation methods, so the
  analyser can resolve relations instead of guessing.
- Suppressed the one verified false-positive class (attribute access on a
  relation resolved to the base Model, which PHPStan cannot see without
  @property docblocks), documented in phpstan.neon.
- Real fixes: cast author_revenue to float before increment (decimal columns
  come back as strings), cast floor() to int for array_slice, drop unnecessary
  nullsafe operators on non-null objects.

Reported issues 99 -> 51. Each remaining item was reviewed individually and is
benign: defensive guards the analyser proves redundant, closure type inference
it cannot follow, and three genuinely unused private helpers left in place.

+++ app/Http/Controllers/Api/Ai/AiBookStudioController.php
@@ -638,7 +638,7 @@ class AiBookStudioController extends Controller
 
         // ALLOCATION LOGIC per requirement (UPDATED FOR EFFICIENCY)
         $totalCapacityWords = $maxPages * 275;
-        $targetWords = floor($totalCapacityWords / $totalSubchapters);
+        $targetWords = (int) floor($totalCapacityWords / $totalSubchapters);
         $pagesPerSubchapter = $targetWords / 275; // For prompt context (can be float now)
 
         // Minimum viable length check (approx 0.5 pages or 130 words)
